commands: move operator command to separate pkg
