+++ command/oper/operator_test.go
@@ -1,0 +1,13 @@
+package oper
+
+import (
+	"strings"
+	"testing"
+)
+
+func TestOperatorCommand_noTabs(t *testing.T) {
+	t.Parallel()
+	if strings.ContainsRune(New().Help(), '\t') {
+		t.Fatal("usage has tabs")
+	}
+}
